Refactor lib-rpc into lib-rpc-core and lib-rpc-app   - agent and conv related RPC methods split out from lib-rpc into lib-rpc-app   - references to lib-rpc crate updated to lib-rpc-app and lib-rpc-core in code and toml   - various lib.rs and mod.rs adjusted
Allows the expected use-case of consumers copying and modifying lib-rpc-app for their domain model  while using the lib-rpc-core as-is.
