+++ crates/libs/lib-rpc-app/src/lib.rs
@@ -1,0 +1,3 @@
+mod rpcs;
+
+pub use rpcs::*;
