+++ crates/libs/lib-rpc-app/src/rpcs/agent_rpc.rs
@@ -1,4 +1,4 @@
-use crate::rpcs::prelude::*;
+use lib_rpc_core::prelude::*;
 use lib_core::model::agent::{
 	Agent, AgentBmc, AgentFilter, AgentForCreate, AgentForUpdate,
 };
