+++ crates/libs/lib-rpc-app/src/rpcs/mod.rs
@@ -1,8 +1,5 @@
 // region:    --- Modules
 
-mod macro_utils;
-mod prelude;
-
 pub mod agent_rpc;
 pub mod conv_rpc;
