+++ crates/libs/lib-rpc-core/src/error.rs
