+++ crates/libs/lib-rpc-core/src/lib.rs
@@ -4,10 +4,9 @@ mod error;
 mod rpc_params;
 mod rpc_result;
 mod rpcs;
+pub mod prelude;
 
 pub use self::error::{Error, Result};
 pub use rpc_params::*;
-
-pub use rpcs::*;
 
 // endregion: --- Modules
