+++ crates/libs/lib-rpc-core/src/rpcs/macro_utils.rs
