+++ crates/libs/lib-rpc-core/src/rpcs/mod.rs
@@ -1,0 +1,4 @@
+// region:    --- Modules
+mod macro_utils;
+
+// endregion: --- Modules
